+++ src/components/Plugins/BusinessCalendar.tsx
@@ -57,7 +57,7 @@ export const BusinessCalendar: React.FC = () => {
           variant="secondary"
           icon="github"
           target="_blank"
-          href="https://github.com/volkovlabs/volkovlabs-calendar-panel"
+          href="https://github.com/volkovlabs/business-calendar"
         >
           GitHub
         </LinkButton>