+++ src/components/Plugins/index.ts
@@ -5,5 +5,6 @@ export * from './BusinessInput';
 export * from './BusinessMedia';
 export * from './BusinessNews';
 export * from './BusinessSatellite';
+export * from './BusinessTable';
 export * from './BusinessText';
 export * from './BusinessVariable';